==> 1. crawling.ipynb <==
import sqlite3
import requests
from bs4 import BeautifulSoup
from tqdm import tqdm

def get_db(db_name):
	db = sqlite3.connect(db_name)
	return db

def close_db(db):
    db.close()

def load_article_id_map(db_name):
    db = get_db(db_name).cursor()
    db.execute(f"""SELECT title, date, origin_link, origin_id
                    FROM article_id_map 
                    WHERE date LIKE '2024%';""")
    rows = db.fetchall()
    return rows

# 테이블 생성 (이미 존재하면 생략)
def create(db_name):
    db = get_db(db_name)
    cursor = db.cursor()
    cursor.execute('''
        CREATE TABLE IF NOT EXISTS article (
            id INTEGER PRIMARY KEY,
            title TEXT NOT NULL,
            date INTEGER NOT NULL,
            content TEXT NOT NULL,
            link TEXT NOT NULL
        )
    ''')
    db.commit()  # 변경사항 저장
    close_db(db) # 접속 종료

def crawling(url):
    try:
        response = requests.get(url)
        response.raise_for_status()  # 요청에 실패하면 예외 발생
        soup = BeautifulSoup(response.text, 'html.parser')
        article_section = soup.find('section', class_='news_view')
        
        if article_section:
            # 광고 등 불필요한 태그를 제거
            for ad in article_section.find_all(['div', 'script', 'figure']):
                ad.decompose()
            
            # 남은 텍스트 추출
            article_text = article_section.get_text(separator=' ', strip=True)
            return article_text
        else:
            return None
    except requests.RequestException as e:
        print(f"HTTP 요청 실패: {e}")
        return None
    
def insert(db_name, rows):
    db = get_db(db_name)
    cursor = db.cursor()
    for row in tqdm(rows):
        content = crawling(row[2])
        if content:
            cursor.execute('''
                INSERT INTO article (id, title, date, content, link)
                VALUES (?, ?, ?, ?, ?)
            ''', (row[3], row[0], row[1], content, row[2])) # origin_id, title, date, content, link
    db.commit()
    close_db(db)

def select(db_name):
    db = get_db(db_name).cursor()
    db.execute(f"""SELECT id, title, date, content, link
                    FROM article;""")
    rows = db.fetchall()
    return rows

rows = load_article_id_map('dataset/pr-db.db')

len(rows)

# 새로운 DB 만들고 insert
create('dataset/2024.db')
insert('dataset/2024.db', rows)

rows = select('dataset/2024.db')
len(rows)

import pandas as pd

df = pd.DataFrame(rows, columns=['id', 'title', 'date', 'content', 'link'])
df.head()

df.to_csv("dataset/2024.csv", index=False)
