==> 2. preprocessing.ipynb <==
### 1. Title, content를 normalize하고 100자 이하의 정보성이 없는 기사는 제거(2024_norm.csv)

from utils.normalize_text import normalize
import hanja
import re

def prerpocessing(sentence):
    # 텍스트 정규화
    sentence = normalize(sentence).replace('  ',' ')
    if re.search(r'[^ㄱ-ㅎ가-힣a-zA-Z0-9…·##μ●▲▶\'!$%&()*+,./:;<=>?@\^`{|}~∼％"\[\]\s-]', sentence):
        sentence = hanja.translate(sentence,'substitution')
    return sentence

import pandas as pd
df = pd.read_csv("dataset/2024.csv")
df.head()

df.drop('link', axis=1, inplace=True)
df.head()

df['title'] = df['title'].apply(prerpocessing)
df['content'] = df['content'].apply(prerpocessing)
df.head()

df.to_csv("dataset/2024_norm.csv", index=False)

### 2. Content를 450자 기준으로 split하여 id_(index) 방식의 새로운 unique_id를 가진 데이터셋으로 변환(2024_norm_split.csv)
- upstage의 임베딩을 사용할 계획인데 512 토큰 이하의 document가 가장 좋은 성능을 보인다고 한다

df = pd.read_csv("dataset/2024_norm.csv")
df.head()

# split한 블록들을 차례대로 누적시켜서 새로 구성할, 빈 DataFrame 생성
df_split = pd.DataFrame(columns=['id', 'title', 'date', 'content'])

# 각 행에 대한 split 작업 수행
for idx in range(len(df)):
    origin_row = df.iloc[idx]
    origin_id = origin_row['id']
    content_length = len(origin_row['content'])
    if content_length > 450: # 450자가 넘어가면 Split 작업 수행
        num_blocks = content_length // 450
        # 450자씩 잘라서 block을 만들고 unique_id를 기존 행의 id로 대체하여 df_split에 추가
        for step in range(num_blocks): # 맨 마지막 블록은 정보성이 많이 떨어질 것으로 판단하여 추가하지 않음
            unique_id = str(origin_id) + '_' + str(step)
            new_row = origin_row.copy()
            new_row['id'] = unique_id
            new_row['content'] = origin_row['content'][(step * 450) : (step + 1) * 450]
            df_split.loc[len(df_split)] = new_row
    else:
        # 바로 df_split에 추가
        unique_id = str(origin_id) + '_0'
        new_row = origin_row.copy()
        new_row['id'] = unique_id
        df_split.loc[len(df_split)] = new_row

print("Before Split :", len(df))
print("After Split :", len(df_split))

df_split.head()

df_split.to_csv("dataset/2024_norm_split.csv", index=False)
