==> 4. Kiwi_customizing.ipynb <==
## 1. 정보성 부족한 기사 제거

import pandas as pd

df = pd.read_csv("dataset/2024_norm_split_ner.csv")
df.head()

length_list = [len(x) for x in df['content'].tolist()]
print(max(length_list), min(length_list)) # 너무 짧은 블록들이 있다

print("under 100 :", len(df[df['content'].str.len() <= 100]))
print("under 150 :", len(df[df['content'].str.len() <= 150]))
print("under 200 :", len(df[df['content'].str.len() <= 200]))
print("under 250 :", len(df[df['content'].str.len() <= 250]))
print("under 300 :", len(df[df['content'].str.len() <= 300]))
print("under 350 :", len(df[df['content'].str.len() <= 350]))
print("under 400 :", len(df[df['content'].str.len() <= 400]))
print("over 400 :", len(df[df['content'].str.len() > 400]))

# 200~250자는 확실히 정보성이 있는 느낌
df[(df['content'].str.len() <= 250) & (df['content'].str.len() >= 200)]['content'].tolist()[:10]

# 150~200자도 정보성이 있어 보인다
df[(df['content'].str.len() <= 200) & (df['content'].str.len() >= 150)]['content'].tolist()[:10]

# 100~150자는 3~4문장 정도 밖에 되지 않으며, 주로 행사 소개, 간단한 소식 전달 수준이므로 150자 이하부터 제거하기로 판단했다.
df[(df['content'].str.len() <= 150) & (df['content'].str.len() >= 100)]['content'].tolist()[:10]

print("짧은 길이의 기사 제거 전 :", len(df))
df = df[df['content'].str.len() >= 150]
print("짧은 길이의 기사 제거 후 :", len(df)) # 약 1200개 정도 제거

df.to_csv("dataset/2024_norm_split_ner_pruning.csv", index=False)

## 2. NER 결과 빈도 분석하여 유효한 Set 추출

import pandas as pd
df = pd.read_csv("dataset/2024_norm_split_ner_pruning.csv")
df.head()

# NER을 전체로 합쳐서 다시 공백 기준으로 나누어 단어 별로 분해
total = ' '.join(df['NER'].tolist()).split() 
len(total) # 총 100만 단어

unique_ner = list(set(total))
len(unique_ner) # 그 중에서 중복 없이 등장한 단어 개수는 약 11만개

from collections import Counter

ner_counts = Counter(total)
ner_counts_df = pd.DataFrame(ner_counts.items(), columns = ['NER', 'Frequency'])

ner_counts_df = ner_counts_df.sort_values(by='Frequency', ascending=False)
len(ner_counts_df)

ner_counts_df.head()

####  빈도 수가 적은 단어들은 NER 성능의 문제 혹은 사용 빈도수가 적은 어휘로 판단하고 사전에 추가하지 않음

ner_counts_df[ner_counts_df['NER'].str.len() == 1]

ner_counts_df = ner_counts_df[ner_counts_df['NER'].str.len() > 1]
len(ner_counts_df)

# 1회만 나온 것은 NER 모델의 오류이거나 거의 사용되지 않는 단어
ner_counts_df[ner_counts_df['Frequency'] == 1].head()

ner_counts_df[ner_counts_df['NER'] == '바른미래당']

# 5번 이상부터는 유의미해 보임
ner_counts_df[ner_counts_df['Frequency'] == 5].head()

len(ner_counts_df[ner_counts_df['Frequency'] >= 5])

- 약 3만개의 기사에서 추출된 NER 약 100만개, 그 중에서 중복없이 나온 단어 총 11만개
- 중복없이 나온 단어들 중에서 빈도 수가 10 미만인 단어들을 제거하여, 유의미하다고 판단한 단어는 약 2만개
- 이 2만개의 단어를 Kiwi 형태소 분석기 사전에 추가하여 BM25의 tokenizer로 사용할 계획

ner_counts_df = ner_counts_df[ner_counts_df['Frequency'] >= 5]
len(ner_counts_df)

ner_counts_df['Frequency'].max()

# 각각의 NER에 우선순위(score)는 빈도값의 역순으로 적용(값이 높을수록 우선순위가 높게 책정됨)
# 빈도값이 높을수록 짧은 경우가 많은데 그런 경우 적게 나온 고유 명사의 일부분에 해당하여 하위 빈도값이 하나의 형태소로 인식되지 않는 문제가 발생
# 따라서 오히려 빈도값의 역순으로 지정하되 값이 0보다 크도록 max+1에서 뺀 값을 적용
# 기존 사전의 경우 대부분의 값이 0이하이며, score를 지정하지 않으면 0이 default로 적용됨

# 추가적으로 NNP로 하는 경우 기존 사전에 있는 단어의 경우 반영이 되지 않는 문제가 발생
# NER에서는 거의 없는 임의의 태그(EC - 연결어미)를 사용해서 강제로 높은 score로 사전에 추가하는 방식에 사용
# (수정) EC로 넣는 경우 어떠한 충돌 때문인지 제대로 작동하지 않음 --> 같은 체언 계열에 NER에는 거의 없는 NNB로 할당하면 문제 없이 작동
with open("model/custom_dict.txt", "w", encoding='utf-8') as f:
    for idx in range(len(ner_counts_df)):
        line = ner_counts_df.iloc[idx]['NER'] + '\t' + 'NNB' + '\t' + str(((ner_counts_df['Frequency'].max() + 1) - ner_counts_df.iloc[idx]['Frequency'])/10000) + '\n'
        f.write(line)

valid_ner_set = ner_counts_df['NER'].tolist()
len(valid_ner_set)

valid_ner_set[:10]

## 3. Kiwi 형태소 분석기 사전 구성

from kiwipiepy import Kiwi
from kiwipiepy.utils import Stopwords

kiwi = Kiwi(num_workers=0, model_type='sbg') # SkipBigram(sbg)
stopwords = Stopwords()

# 사전에 '바른미래당'이 없어서 분리되는 것을 확인 가능
print(kiwi.tokenize("바른미래당 후보가 밥을 먹었다",stopwords=stopwords))

# 카카오프렌즈의 경우 '카카오', '프렌즈'의 score가 더 높아서 하나의 단어로 인식되지 않는 문제가 있다
print(kiwi.tokenize("카카오프렌즈를 잘 이용하고 있다",stopwords=stopwords))

# 18887개 중 5개를 제외하고는 다 새롭게 사전에 추가된 것을 확인
# 기존의 사전은 https://raw.githubusercontent.com/bab2min/Kiwi/main/ModelGenerator/default.dict에서 확인 가능
# Documentation은 https://github.com/bab2min/kiwipiepy/blob/main/README.md를 참고
kiwi.load_user_dictionary(dict_path="model/custom_dict.txt")

kiwi.tokenize("바른미래당 후보가 밥을 먹었다",stopwords=stopwords)

kiwi.tokenize("카카오프렌즈를 잘 이용하고 있다",stopwords=stopwords)

import pandas as pd

df = pd.read_csv("dataset/2024_norm_split.csv")
df.head()

corpus = df['content'].tolist()
len(corpus)

from tqdm import tqdm

# 전체 데이터셋에 대해서도 오류없이 tokenizing이 잘 되는 모습을 확인
for content in tqdm(corpus):
    try:
        kiwi.tokenize(content,stopwords=stopwords)
    except Exception as e:
        print(e)
