==> 5. BM25_model.ipynb <==
## Corpus 구성

import pandas as pd

df = pd.read_csv("dataset/2024_norm_split.csv")
df.head()

corpus = df['content'].tolist()
len(corpus)

corpus[:10]

## Custom Kiwi 형태소 분석기로 BM25용 Analyzer 생성

from kiwipiepy import Kiwi
from kiwipiepy.utils import Stopwords
from typing import List

class Kiwi_Tokenizer:
    def __init__(self, dict_path : str):
        self.kiwi = Kiwi(num_workers=0, model_type='sbg') # SkipBigram(sbg)
        self.stopwords = Stopwords()
        self.kiwi.load_user_dictionary(dict_path=dict_path)

    def __call__(self, sent : str) -> List[str]:
        tokens = self.kiwi.tokenize(sent, stopwords=self.stopwords)
        return [token.form for token in tokens]

my_tokenizer = Kiwi_Tokenizer("model/custom_dict.txt")

tokens = my_tokenizer(corpus[0])
print(tokens)

## BM25 모델 생성

from langchain_milvus.utils.sparse import BM25SparseEmbedding
from pymilvus.model.sparse import BM25EmbeddingFunction

class Custom_BM25(BM25SparseEmbedding):
    def __init__(self, corpus: List[str], tokenizer):
        # super().__init__(corpus=corpus) # fit도 자동으로 하기 때문에 이를 분리하기 위해 __init__은 overriding
        self.corpus = corpus
        self.analyzer = tokenizer # default로 설정된 build_default_analyer가 아닌, 입력된 tokenizer를 analyzer로 설정
        self.bm25_ef = BM25EmbeddingFunction(self.analyzer, num_workers=1)
    
    # corpus에 대해 analyzer를 이용하여 학습하여 BM25 모델을 생성하는 메소드
    def fit(self):
        self.bm25_ef.fit(self.corpus)
    
    # corpus에 대해 fit된 모델을 저장하는 메소드
    def save(self, filename : str):
        self.bm25_ef.save(filename)
    
    # fit된 모델을 로드하는 메소드
    def load(self, filename : str):
        self.bm25_ef.load(filename)
    
    # 그대로 사용하는 메소드
    # embed_query(self, text: str) -> Dict[int, float]
    # embed_documents(self, texts: List[str]) -> List[Dict[int, float]]

# corpus와 analyzer가 설정된 상태까지
my_BM25 = Custom_BM25(corpus=corpus, tokenizer=my_tokenizer)

# 선언 당시 저장된 corpus와 analyzer를 기반으로 BM25 모델을 학습
my_BM25.fit() # 총 2시간 소요

# 학습이 완료된 BM25 모델을 저장
my_BM25.save("model/bm25_model.json")

## Sparse Embedding

# 토크나이저, BM25 모델 로드 등 초기 세팅
tokenizer = Kiwi_Tokenizer("model/custom_dict.txt")
custom_bm25 = Custom_BM25(corpus=[], tokenizer=tokenizer)
custom_bm25.load("model/bm25_model.json")

import pandas as pd

df = pd.read_csv("dataset/2024_norm_split_ner_pruning.csv")
df.head()

docs_embedding = custom_bm25.embed_documents(df['content'].tolist()[:10])
for doc_embedding in docs_embedding:
    print(doc_embedding)
