==> 8. document_input_compare.ipynb <==
## Query가 아닌 Document 형태로 input이 될 경우

from pymilvus import connections, utility, FieldSchema, CollectionSchema, DataType, Collection

QUERY_MODEL = 'solar-embedding-1-large-query'
PASSAGE_MODEL = 'solar-embedding-1-large-passage'

DENSE_FIELD = "dense_vector"
DENSE_DIMENSION = 4096 
DENSE_INDEX = {"index_type" : "FLAT", "metric_type" : "IP"}

SPARSE_FILED = "sparse_vector"
SPARSE_INDEX = {"index_type" : "SPARSE_INVERTED_INDEX", "metric_type" : "IP"}

# Collection 접속 및 생성
def milvus_connect(collection_name):
    # Milvus 설정
    MILVUS_HOST = '127.0.0.1'
    MILVUS_PORT = '19530'     

    # Milvus 연결
    connections.connect(host=MILVUS_HOST, port=MILVUS_PORT)
    if connections:
        print("Milvus connected")
    else:
        exit()

    # 컬렉션 존재 여부 확인 및 생성
    if utility.has_collection(collection_name):
        collection = Collection(collection_name)
        print(f"Collection '{collection_name}' loaded.")
    else:
        print(f"Collection '{collection_name}' does not exist.")
        print(f"Create '{collection_name}' Collection.")
        
        field_args = [
            FieldSchema(name='id', dtype=DataType.VARCHAR, is_primary=True, max_length=100),        # split된 block의 고유 ID
            FieldSchema(name='title', dtype=DataType.VARCHAR, max_length=2000),
            FieldSchema(name='date', dtype=DataType.INT64),
            FieldSchema(name='content', dtype=DataType.VARCHAR, max_length=5000),
            FieldSchema(name='NER', dtype=DataType.VARCHAR, max_length=3000),
            FieldSchema(name=DENSE_FIELD, dtype=DataType.FLOAT_VECTOR, dim=DENSE_DIMENSION),
            FieldSchema(name=SPARSE_FILED, dtype=DataType.SPARSE_FLOAT_VECTOR)
        ]
        
        # 스키마 정의
        schema = CollectionSchema(fields=field_args)
        
        # 컬렉션 생성
        collection = Collection(name=collection_name, schema=schema)
        
        # 인덱스 생성
        collection.create_index("dense_vector", DENSE_INDEX)
        collection.create_index("sparse_vector", SPARSE_INDEX)

        collection.flush()

        print(f"Collection '{collection_name}' created and loaded.")
    # 생성된 컬렉션 반환
    return collection

# Collection 제거용
def drop_collection(collection_name):
    MILVUS_HOST = '127.0.0.1'
    MILVUS_PORT = '19530' 
    connections.connect(host=MILVUS_HOST, port=MILVUS_PORT)
    if utility.has_collection(collection_name):
        utility.drop_collection(collection_name)
        print(f"Drop {collection_name} Complete.")
    else:
        print(f"Can't find {collection_name}.")

donga = milvus_connect("Donga")
print(donga)

donga_800 = milvus_connect("Donga_800")
print(donga_800)

## Donga 세팅

# Sparse Embedding : Custom BM25
from model.bm25_model import Custom_BM25, Kiwi_Tokenizer

tokenizer = Kiwi_Tokenizer("model/custom_dict.txt")
custom_bm25 = Custom_BM25(corpus=[], tokenizer=tokenizer)
custom_bm25.load("model/bm25_model.json")
print(custom_bm25)

# Dense Embedding : Upstage/solar-embedding-1-large
from langchain_upstage import UpstageEmbeddings
from dotenv import load_dotenv
import os

load_dotenv()

dense_embedding_func = UpstageEmbeddings(
    model="solar-embedding-1-large-passage",
    upstage_api_key=os.getenv("UPSTAGE_API_KEY")
)

from langchain_milvus.retrievers import MilvusCollectionHybridSearchRetriever
from pymilvus import WeightedRanker

def hybrid_search(query : str, filter : str = None):
    sparse_search_params = {"metric_type": "IP"}
    dense_search_params = {"metric_type": "IP", "params": {}}

    params = {
        "collection" : donga,
        "rerank" : WeightedRanker(0.7, 0.3),
        "anns_fields" : [DENSE_FIELD, SPARSE_FILED],
        "field_embeddings" : [dense_embedding_func, custom_bm25],
        "field_search_params" : [dense_search_params, sparse_search_params],
        "top_k" : 20,
        "text_field" : 'content',
    }
    if filter:
        params['field_exprs'] = [filter, filter] # 검색 조건 추가(dense와 sparse에 동일하게 적용)
    retriever = MilvusCollectionHybridSearchRetriever(**params)
    
    retrieved_docs = retriever.invoke(query)
    data = []
    unique_id = []
    for retreived_doc in retrieved_docs:
        doc_id = retreived_doc.metadata['id'].split("_")[0] # 기사 고유 ID 추출
        if doc_id not in unique_id: # 해당 기사가 아직 반환 리스트에 없는 경우만
            data.append(retreived_doc)
            unique_id.append(doc_id)
        if len(data) == 5: break # 상위 5개만 반환
    return data

## Donga_800 세팅

from openai import OpenAI
import numpy as np

INDEX_PARAMS = {
    'metric_type': 'COSINE',     # "L2", "IP", "COSINE" 
    'index_type': "IVF_FLAT",    # "IVF_FLAT","IVF_SQ8", "IVF_PQ", "HNSW", "ANNOY"
    'params': {"nlist": 128},
}

client = OpenAI(
    api_key = os.getenv("UPSTAGE_API_KEY"),
    base_url = "https://api.upstage.ai/v1/solar"
)

def vector_search(collection, query, index_params, option=None):
    # Query embedding 생성
    query_embedding = client.embeddings.create(
        model="solar-embedding-1-large-passage",
        input=query
    ).data[0].embedding
    query_embedding = np.asarray(query_embedding)
    
    # 공통 search 파라미터 설정
    search_params = {
        "data": [query_embedding],
        "anns_field": 'embedding',
        "param": index_params,
        "limit": 5,
        "output_fields": ['title', 'date', 'content', 'link']
    }

    # 옵션이 있는 경우 expr 추가
    if option:
        search_params["expr"] = option

    # 검색 실행
    result = collection.search(**search_params)

    # 결과 처리
    result_list = []
    for hits in result:
        for hit in hits:
            result_list.append(hit.to_dict())
    
    return result_list

# Test

# 출처 기사 : 1평에 24명 ‘빽빽’… “미등록 공연장, 제2 이태원 참사 우려”, 2024.08.02 (https://n.news.naver.com/mnews/article/020/0003579874)
# 원문 기사는 1800자 정도, 이를 chatgpt 4o를 이용하여 700 ~ 800자 분량으로 요약
passage1 = """
지난달 29일 서울 성동구 복합문화공간 에스팩토리에서 압사 우려로 공연이 중단됐으며, 당시 1㎡당 최대 7명이 밀집해 있었습니다. 이는 2022년 10월 이태원 참사 당시 밀집도에 근접하는 수준이었습니다. 에스팩토리는 공연법상 등록되지 않은 공연장으로, 수용 인원 규제가 없어 관리 사각지대에 놓여 있었습니다. 전문가들은 공연장 3층의 밀집도를 분석한 결과, 1인당 점용 면적이 손바닥 크기인 0.14∼0.25㎡에 불과했다고 밝혔습니다. 현행법상 바닥 면적이 0.5㎡ 이하면 위험한 상태로 분류됩니다. 에스팩토리의 공연 주관사 측은 자체 기준으로 관객 수를 정했다고 밝혔지만, 전문가들은 이 같은 미등록 공연장에서 대형 사고가 발생할 수 있다고 경고합니다. 향후 인천 송도 록페스티벌, 워터밤 등 대규모 공연들이 예정되어 있어 대책 마련이 시급합니다. 전문가들은 공연장 내 수용 인원을 안전과 직결된 문제로 보고, 법적 규정을 만들어야 한다고 주장합니다. 공연 도중 인파가 무대 쪽으로 몰리는 현상과 안전요원의 지시가 통하지 않는 상황이 반복될 수 있어 철저한 인파 관리 대책이 필요합니다. 지난 6월 강원에서 열린 페스티벌에서도 무대 근처 스탠딩 구역에서 관객들이 압박을 받아 어려움을 겪었으며, 안전요원의 지시가 잘 통제되지 않았습니다. 이러한 문제를 해결하기 위해 공연장의 모든 구역 내 상시 안전 공간을 확보하고 세심한 관리 대책을 마련해야 한다는 지적이 나옵니다.
"""

hybrid_search(passage1, filter='date >= 20240601') # donga_800과 동일한 데이터 기반으로 하기 위해 6월~7월 기사만으로 구성

vector_search(donga_800, passage1, INDEX_PARAMS)

hybrid_search(passage1) # 필터없이 2024년 전체 기사를 기반으로

# 출처 기사 : 끝이 보이는 고금리 시대… 美연준, 9월 인하 시사, 2024.08.02 (https://n.news.naver.com/mnews/article/020/0003579902)
# 원문 기사는 4000자 정도, 이를 chatgpt 4o를 이용하여 700 ~ 800자 분량으로 요약
passage2 = """
미국 연방준비제도(Fed)가 9월 금리 인하를 강력히 시사하면서 글로벌 고금리 사이클이 2년 반 만에 마무리될 전망이다. 지난달 31일 연준은 연방공개시장위원회(FOMC)에서 기준금리를 동결했으나, 제롬 파월 의장은 9월 회의에서 금리 인하가 논의될 수 있다고 밝혔다. 시장은 9월 금리 인하 가능성을 거의 100%로 보고 있으며, 물가 안정과 고용시장 둔화가 그 배경이다. 6월 개인소비지출(PCE) 상승률은 전년 대비 2.5%로 안정됐고, 실업률은 4.1%로 상승했다. 미국 외 주요국들도 금리 인하에 나서고 있다. 캐나다는 두 달 연속 금리를 인하했고, 유럽중앙은행(ECB)과 중국, 영국도 금리를 낮췄다. 이러한 금리 인하는 글로벌 경제에 큰 변화를 가져올 전망이다. 금융시장에서는 자산 가치가 상승 압력을 받을 수 있고, 민간 소비와 기업 투자도 늘어날 가능성이 있다. 강달러 현상이 완화되면서 아시아 등 다른 나라의 통화 가치가 반등할 여지도 있다. 한국은행도 금리 인하 가능성을 검토 중이다. 내수 경기 침체와 물가 상승세 둔화로 한은의 10월 금리 인하설이 힘을 얻고 있다. 이창용 한은 총재는 금리 인하 논의를 시작했다고 밝혔다. 한은은 2021년 8월부터 금리를 인상해 현재 3.50%를 유지해왔으며, 고금리로 인해 가계 이자상환 부담이 늘고 내수와 부동산 경기가 둔화되는 등 부작용이 있었다. 정부와 여당도 한은에 금리 인하를 요구하고 있다. 그러나 금리 인하가 부동산 시장을 자극할 수 있다는 점은 한은에 부담이다. 부동산값 상승이 은행권 주택담보대출을 늘려 가계부채가 증가할 수 있기 때문이다. 미국 대선과 중동 전쟁, 달러화 강세, 국제 유가 급등 등으로 국내 물가가 다시 불안해질 가능성도 있다. 전문가들은 한은이 10월 금리 인하에 나설 가능성을 예상하면서도, 부동산 대책 효과 등을 살핀 뒤 결정할 것으로 본다.
"""

hybrid_search(passage2, filter='date >= 20240601') # donga_800과 동일한 데이터 기반으로 하기 위해 6월~7월 기사만으로 구성

vector_search(donga_800, passage2, INDEX_PARAMS)

hybrid_search(passage2) # 필터없이 2024년 전체 기사를 기반으로
